+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Revamp of the progress bar: static regardless of screen, turns green when part is completed, resets when returning to main questionnaire.
+- Removed final semicolon \(;\) from the multiple choice, multiple answer question type.
+- Removed AttributeError from MultiMulipleChoiceQQuestion, because it was not relevant.
 
 ### Deprecated